refactor: 유틸리티 함수 생성(fetchData)
